+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Version 1.2.9
 
- * 
+ * Add osm\_long\_key to badges to make getting badge stock easier
 
 ## Version 1.2.8